Search GST on real-time basis and not within the saved data
As of now, if user is searching for GST number, existing tables are being searched for firstly and if GST number is not found in database, then API is called to fetch data from GST service provider. This should not be done for search through this form.

Refer ST-1877 for more details.